requirements.txt
djangorestframework==3.14.0
tensorflow>=2.6.0
opencv-python>=4.5.0
celery>=5.0.0
redis>=3.5.0
numpy>=1.19.0